macroquad

+++ src/food.rs
@@ -10,8 +10,8 @@ impl Food {
     pub fn create_new(&mut self) {
         let mut rng = rand::rng();
         self.count += 1;
-        let new_x = (rng.random_range(0..60) * crate::STEP) as f64;
-        let new_y = (rng.random_range(0..60) * crate::STEP) as f64;
+        let new_x = (rng.random_range(0..60) * crate::STEP) as f32;
+        let new_y = (rng.random_range(0..60) * crate::STEP) as f32;
         self.pos = Position::new(new_x, new_y);
     }
 }
